simplified all the workers trying to update the screen into a single worker to resolve issue with vsync
